[MIG] Migrated purchase_exception to v12.
