+++ purchase_exception/README.rst
@@ -18,12 +18,11 @@ Usage
 
 .. image:: https://odoo-community.org/website/image/ir.attachment/5784_f2813bd/datas
    :alt: Try me on Runbot
-   :target: https://runbot.odoo-community.org/runbot/167/11.0
+   :target: https://runbot.odoo-community.org/runbot/167/12.0
 
 
 Bug Tracker
 ===========
-
 
 Bugs are tracked on `GitHub Issues
 <https://github.com/OCA/purchase-workflow/issues>`_. In case of trouble, please
@@ -39,6 +38,7 @@ Contributors
 ------------
 
 * Mourad EL HADJ MIMOUNE <mourad.elhadj.mimoune@akretion.com>
+* Sudhir Arya <sudhir@erpharbor.com>
 
 Maintainer
 ----------
