+++ purchase_exception/__init__.py
@@ -1,2 +1,2 @@
-
+# -*- coding: utf-8 -*-
 from . import models, wizard
