+++ purchase_exception/models/__init__.py
@@ -1,2 +1,3 @@
 
+from . import exception_rule
 from . import purchase
