+++ purchase_exception/wizard/__init__.py
@@ -1,2 +1,2 @@
-
+# -*- coding: utf-8 -*-
 from . import purchase_exception_confirm
