+++ purchase_exception/wizard/purchase_exception_confirm_view.xml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
+
     <record id="action_purchase_exception_confirm" model="ir.actions.act_window">
         <field name="name">Outstanding exceptions to manage</field>
         <field name="type">ir.actions.act_window</field>
@@ -9,4 +10,5 @@
         <field name="view_id" ref="base_exception.view_exception_rule_confirm"/>
         <field name="target">new</field>
     </record>
+
 </odoo>
